+++ src/pudl/output/pudltabl.py
@@ -221,7 +221,7 @@ class PudlTabl:
             # plant parts
             "out_eia__yearly_generators_by_ownership": "gens_mega_eia",
             "out_eia__yearly_plant_parts": "plant_parts_eia",
-            "out__yearly_plants_all_ferc1_plant_parts_eia": "ferc1_eia",
+            "out_pudl__yearly_assn_eia_ferc1_plant_parts": "ferc1_eia",
         }
 
         table_method_map_any_agg = {